Changes as suggested
