+++ src/pages/CommandIndex/CommandIndex.tsx
@@ -4,7 +4,7 @@ import { Divider } from 'components/Divider'
 import { PageHeader } from 'components/PageHeader'
 import { Table } from 'components/Table'
 import { PermissionsContainer } from 'containers/PermissionsContainer'
-import { useCommandsParameterized } from 'hooks/useCommandParameterized'
+import { useCommandsParameterized } from 'hooks/useCommandsParameterized'
 import { commandsFromSystems } from 'hooks/useCommands'
 import { useCommandIndexTableColumns } from 'pages/CommandIndex'
 import { useEffect, useState } from 'react'
